Make app router pages smaller by reducing translations size in global 404 page
