+++ apps/store/src/app/not-found.tsx
@@ -1,15 +1,22 @@
-import { NotFoundPageContent } from '@/appComponents/RootLayout/NotFoundPageContent' // Fallback to se/not-found
-import { FALLBACK_LOCALE } from '@/utils/l10n/locales'
+import { initTranslations } from '@/app/i18n'
+import { TranslationsProvider } from '@/appComponents/providers/TranslationsProvider'
+import { NotFoundPageContent } from '@/appComponents/RootLayout/NotFoundPageContent'
+import { StoreLayout } from '@/appComponents/StoreLayout'
+import { FALLBACK_LOCALE } from '@/utils/l10n/locales' // Fallback to se/not-found
 import { toRoutingLocale } from '@/utils/l10n/localeUtils'
-import LocaleLayout from './[locale]/layout'
 
 // Global fallback, has to be located at the top of app dir
 // Hence manual import and usage of layout
 // We're doing locale-specific 404s at [locale]/not-found
-export default function NotFound() {
+export default async function RootNotFoundPage() {
+  const locale = toRoutingLocale(FALLBACK_LOCALE)
+  // Only loading common namespace since this page's React tree is added to every normal page - size matters
+  const { resources } = await initTranslations(locale, { ns: ['common'] })
   return (
-    <LocaleLayout params={{ locale: toRoutingLocale(FALLBACK_LOCALE) }}>
-      <NotFoundPageContent />
-    </LocaleLayout>
+    <TranslationsProvider locale={locale} resources={resources}>
+      <StoreLayout locale={locale}>
+        <NotFoundPageContent />
+      </StoreLayout>
+    </TranslationsProvider>
   )
 }
